+++ app/controllers/devise/unlocks_controller.rb
@@ -14,7 +14,7 @@ class Devise::UnlocksController < DeviseController
     if successfully_sent?(resource)
       respond_with({}, location: after_sending_unlock_instructions_path_for(resource))
     else
-      respond_with(resource)
+      respond_with(resource, location: new_unlock_path(resource_name))
     end
   end